+++ Azure/connectors/azure_blob.py
@@ -3,11 +3,10 @@
 import asyncio
 import time
 from datetime import datetime, timedelta, timezone
-from typing import Any, Optional
 from gzip import decompress
+from typing import Any, Optional
 
 import aiofiles
-import orjson
 from dateutil.parser import isoparse
 from loguru import logger
 from pydantic import Field
@@ -17,8 +16,8 @@ from sekoia_automation.connector import DefaultConnectorConfiguration
 from sekoia_automation.module import Module
 from sekoia_automation.storage import PersistentJSON
 
-from azure_helpers.storage import AzureBlobStorageConfig, AzureBlobStorageWrapper
 from azure_helpers.io import is_gzip_compressed
+from azure_helpers.storage import AzureBlobStorageConfig, AzureBlobStorageWrapper
 
 from .metrics import EVENTS_LAG, FORWARD_EVENTS_DURATION, OUTCOMING_EVENTS
 
@@ -101,10 +100,7 @@ class AzureBlobConnector(AsyncConnector):
                 if _last_modified_date is None or blob.last_modified > _last_modified_date:
                     _last_modified_date = blob.last_modified
 
-                file, content = (
-                    # TODO: Check if we want to avoid downloading the file to local machine
-                    await self.azure_blob_wrapper().download_blob(blob.name, download=True)
-                )
+                file, content = await self.azure_blob_wrapper().download_blob(blob.name, download=True)
 
                 if file:
                     async with aiofiles.open(file, "rb") as file_data:
@@ -113,16 +109,14 @@ class AzureBlobConnector(AsyncConnector):
                         if is_gzip_compressed(file_content):
                             file_content = decompress(file_content)
 
-                        records.extend(
-                            self.format_blob_data(orjson.loads(file_content.decode("utf-8")), self.last_event_date)
-                        )
+                        records.extend(file_content.decode("utf-8").split("\n"))
 
                     await delete_file(file)
 
                 if content:
-                    records.extend(self.format_blob_data(orjson.loads(content), self.last_event_date))
+                    records.extend(content.decode("utf-8").split("\n"))
 
-        result: list[str] = await self.push_data_to_intakes([orjson.dumps(event).decode("utf-8") for event in records])
+        result: list[str] = await self.push_data_to_intakes(events=records)
 
         with self.context as cache:
             logger.info(
@@ -133,47 +127,6 @@ class AzureBlobConnector(AsyncConnector):
             cache["last_event_date"] = _last_modified_date.isoformat()
 
         return result
-
-    @staticmethod
-    def format_blob_data(data: dict[Any, Any], time_filter: datetime | None) -> list[dict[Any, Any]]:
-        """
-        Format blob data.
-
-        Main purpose of this function is to format input data to supported intake format:
-            https://learn.microsoft.com/en-us/azure/network-watcher/vnet-flow-logs-overview
-
-        Args:
-            data: dict[Any, Any]
-            time_filter: datetime | None
-
-        Returns:
-            list[dict[Any, Any]]:
-        """
-        modified_result = []
-        for line in data.get("records", []):
-            line_time = isoparse(line["time"]).astimezone(timezone.utc)
-
-            # If the record is too old, ignore it.
-            if time_filter and line_time < time_filter:
-                continue
-
-            result = {
-                "macAddress": line["macAddress"],
-                "operationName": line["operationName"],
-                "resourceId": line["resourceId"],
-                "time": line["time"],
-            }
-
-            for property_flow in line["properties"]["flows"]:
-                result["rule"] = property_flow["rule"]
-
-                for flow in property_flow["flows"]:
-                    for flow_tuple in flow["flowTuples"]:
-                        result["flow.0"] = flow_tuple
-
-            modified_result.append(result)
-
-        return modified_result
 
     def run(self) -> None:  # pragma: no cover
         """Runs Azure Blob Storage."""
@@ -187,7 +140,7 @@ class AzureBlobConnector(AsyncConnector):
                     processing_start = time.time()
                     if previous_processing_end is not None:
                         EVENTS_LAG.labels(intake_key=self.configuration.intake_key).set(
-                            processing_start - previous_processing_end
+                            processing_start - self.last_event_date.timestamp()
                         )
 
                     message_ids: list[str] = loop.run_until_complete(self.get_azure_blob_data())
@@ -198,9 +151,8 @@ class AzureBlobConnector(AsyncConnector):
                     if len(message_ids) > 0:
                         log_message = "Pushed {0} records".format(len(message_ids))
 
-                    logger.info(log_message)
                     self.log(message=log_message, level="info")
-                    logger.info(log_message)
+
                     logger.info(
                         "Processing took {processing_time} seconds",
                         processing_time=(processing_end - processing_start),